continued polish
